+++ Changelog.md
@@ -34,6 +34,7 @@ Added a Markdown Preview button to the `TextViewer` component, shown only for Ma
 - Added variable to enable simplecov in `spec_helper.rb` if and only if COVERAGE=true (#7960)
 - Added missing foreign keys in seed data (#8006)
 - Added tests for `AnnotationCategory` to cover an error when uploading a CSV for `add_by_row` (#8177)
+- Added tests for timed and untimed extensions returned by `GroupsController#index` (#8180)
 
 ## [v2.10.4]